turretstats UI equipment slot correct display
The turretstatsUI now shows the correct equipment for each turret. Included a UpdateUI call when the turretstatsUI is hidden (i.e. when it is turned of because of building or when a different turret is selected). This ensures that the UI is only updated to the currently selected turret when changing UI (i.e. hiding and displaying new turrets UI). The OnItemChanged += UpdateUI is still kept so that when equiping a new item, it is immediately displayed and doesnt need an extra hide->display action.