Upgrading turret keeps the support skills
saved the previous turrets equipment into temporaryEquipment variable. Then upgrade turret and equip() all the items in the temporaryEquipment variable to the upgraded turret. From the turret script, Start() is changed to Awake() as instantiating the upgraded turret Awake() (but not Start()) is called to set the number of eqipment slots to 2 so that the equip() right after instantiating can be successfully executed.