Highlighting button
TurretUpgrade and Sell are highlighted when hover over them. Same with skill gems. EquipmentSlots now do not change color when hover over them. This is so that people dont get the idea that these are buttons (selling equipment probably wont be implemented).

The node renderer has been changed from private to public, so other scripts can access it. Need to somehow be able to keep the node (and selected shop turret) highlighted the whole time while they are selected.