Added Highest wave reached
Added a "highscore" to each level in the level selector, indicating the highest wave reached. Highest wave reached is set in playerprefs during the EXP calculation in GameManager.GainLevelEXP(), because it is only saved when the game is over and not when someone just quits through the menu. The name under which the highest wave is reached is the SceneName + "_Max". So the scene name is called dynamically.

Debug: When the last enemy also removes the last life, the BonusItemUI is called before the GameIsOver = true is set. Now the BonusItemUI is set to false when PlayerStats.Lives <= 0.