Enemy movement Update
Changed the buffer distance at which the next wavepoint is chosen to be dynamic, based on enemy movement speed. Which is (enemy.speed / 25). This makes that slower enemies, which are often larger, get closer to the wavepoint (center of the road) before turning towards the next wavepoint. While smaller and faster enemies have larger buffer to circumvent circulating a wavepoint. This also means that larger enemies have a lower chance to overlay firepoints of the standard turret.