Added Range indicator
A range indicator prefab was made by creating a cylinder and setting its height to 0.00001. This gives a circle. The rendering mode is set to transparent.
This prefab is then instantiated with the nodeUI (also written in the NodeUI) whenever the UpdateUI() is called, so its appearance can dynamically change when the stats change.
The prefab is destroyed whenever the UpdateUI is called again (in case of equiping items) and when the nodeUI.Hide() is called.