Freeze function
Added FreezeSeconds attribute to towers.

Added freeze function to EnemyMovement, where it sets speed to 0 and starts a coroutine, when the Enemy.froze > 0. Sets Enemy.froze back to 0 at the end of the coroutine. Enemy.froze is set during the Damage() function, to the towers freezeSeconds status (currently this is 0.5 for the frost tower and 0 for all other towers, meaning only frost tower can freeze). The speed is set back to basespeed at the end of the Update() loop.

Also added Enemy.froze setting in the Damage() function of the bullet.sc and set the bullet.freezeSeconds to the turrets freezeseconds after the bullet is instantiated, so bullets can also freeze if needed.

This status cannot be overwritten or updated as the coroutine isnt stopped or updated in the middle.

Reduced Frost tower fire rate, to adjust for their freeze effect.

Update Frost tower tooltip.