Randomized BonusItem
Created RoundBonusItemUI.cs and linked it tot the ROundBonusItemUI gameObject in the OverlayCanvas. New items must be added to that list for them to be available. The script, upon awakening, for each active slot, it randomly takes an item from the itemlist and overrides the icon (sprite) and adds the item to a onClick() event function whereby it will be added to the inventory.

As such, removed the onlick event of these buttons from the inspector.