Bug: bullet no movement when target == null, bonusItem only rerolls on new game.
Bug 1:
Create a second direction (Vector3 dir2), which saves the direction the turret (partToRotate) is facing. When the bullet is instantiated and target == null, dir2 takes over dir so that the bullet has a direction to move towards.

bug 2:
created a seperate function shuffleBonusItem() which does all the randomizing and putting items on the buttons. This function is called whenever the BonusItemUI is deactivated (void OnDisable).

Somehow, repeatedly calling onClick.AddListener() echos the previous instructions (i.e. 2 then 3 then 4 items are added with each click. So Before everything is set,  onClick.RemoveAllListeners() is called to clean the button off onclick events.