Added Artillery Tower
Added Artillery tower which shoots slow projectile (shell) on a target area (position is where the enemy was spotted). The shell explodes on impact dealing AOE dmg.

Reduced the number of bosses from 5 to 3 at the "last" level.

Changed the way bullet behaviour is determined in the script. Instead of switching based on explosionRadius, now the projectile type is to bool and switched accordingly.