Added Invincible enemy variation. Bug fixes
Added invincible enemies, who become invincible for x seconds after their health drop to less than half (hard coded in Enemy.cs). This triggers a coroutine, which makes enables their halo emissing a yellow light. After x seconds the halo disappears and invincibilityTime for the enemy is set to 0, so that it can never become invincible again. (this is important for regen enemies.

Fixed a Bug where enemies with special properties do not take damage from projectiles. This is because the accessories on the enemies "absorb" the impact. Disable the collider of the accessories fixed the issue.

Removed unused materials

reduced regen rate of tough enemies from 0.2 to 0.1.