Balancing and bugfix
Rebalanced tower damage and cost
reduced tough enemy regen rate from 0.1 to 0.05

Bug fix: standard turret firepoint still not enough. Moved the head and firepoint to the start of the barrel.

bug fix: second shot from artillery shoots to the start point. Create a variable that saves the location of the last known enemy when shoot() is called.