Added Burning Skill
Added a Burning skill that applies a debuff on the enemy that burns it for 3 seconds (coded in the coroutine "for" loop 6 * 0.5s ). The source calls the CallingBurnFromEnemy(); function from enemy.cs script, which then in turn calls stop- and startCoroutine("burn"), which starts to damage the enemy for 0.25 of the source hit damage every 0.5s (= 50% damage per second). The coroutine has to be called from within enemy.cs as otherwise the coroutine stops as soon as the bullet is destroyed.

Also, forgot to mention is that Crit is determined under the Damage() function, thus for AOE each damage instance is calculated seperately.

Still need to add skillbonus for burning.