Added RangeIndicator when building
Added the same RangeIndicator when hovering over free nodes to indicate the range of the tower so that players know the target  aquisition area before deciding to build a tower.
Using the BuildManager.cs when a tower is selected to be build using the SelectTurretToBuild() function, turretRange is set to the baseRange of the turret which is pulled from the turret script of the prefab (turret.prefab.GetComponent<turret>().baserange). This buildManager.turretRange float variable is then pulled by the node.cs script when OnMouseEnter() to transform.localScale the Instantiated _rangeIndicator GameObject.

_rangeIndicator is destroyed OnMouseExit().

Bugfix levelToUnlock when completing a level.