Added "How to play"
Added annotated screenshots of How to Play. Screenshots are images converted and saved into a Sprits[] array. SliderPanel and ImagePanel are added. ImagePanel.image changes based on the order of the image in the array whenever the buttons are pressed.

PanelOpener.cs contains the instruction to slide the panel into view and out of view.

TutorialPages.cs contains the information of the images and mechanism of changing these on button press.